pyproject.toml
[project]
name = "not-your-it-guy"
version = "0.1.0"
description = "OpenAI-compatible agent backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "openai>=1.30.0",
    "langgraph>=0.3.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1.0",
    "sqlalchemy-utils>=0.41.0",
    "loguru>=0.7.3",
]

[project.scripts]
serve = "not_your_it_guy.main:main"
start = "not_your_it_guy.main:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/not_your_it_guy"]

